pyproject.toml
[project]
name = "surgitrack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "docx2txt>=0.9",
    "fastapi>=0.116.0",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-google-genai>=2.1.7",
    "langchain-ollama>=0.3.4",
    "manim>=0.18.1",
    "moviepy==1.0.3",
    "numpy>=1.24.0",
    "ollama>=0.5.1",
    "opencv-python>=4.8.0",
    "openai>=1.94.0",
    "pandas>=2.3.1",
    "pillow>=10.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "spotipy>=2.25.1",
    "uvicorn[standard]>=0.35.0",
    "matplotlib>=3.10.3",
    "ultralytics>=8.3.169",
    "roboflow>=1.2.2",
    "boto3>=1.39.15",
]
